% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_caviar.R
\name{RollCAViaR}
\alias{RollCAViaR}
\title{Rolling window forecast of VaR and ES using CAViaR}
\usage{
RollCAViaR(
  data,
  c,
  n,
  m,
  r = 10,
  var_model = c("ADAPTIVE", "SAV", "AS", "indirectGARCH"),
  es_model = c("MULT", "AR"),
  loss = "AL",
  lb = -5,
  ub = 5,
  grid_size = 10000,
  verbose = 0,
  ...
)
}
\description{
Rolling window forecast of VaR and ES using CAViaR
}
